7.11. Добро пожаловать, или посторонним вход воспрещён (часть 1) v2
