+++ project/src/pages/add-review/add-review.tsx
@@ -1,6 +1,7 @@
 import { Link, useParams } from 'react-router-dom';
 import AddReviewComment from '../../components/add-review-comment/add-review-comment';
 import Logo from '../../components/logo/logo';
+import SignOutComponent from '../../components/sign-out-component/sign-out-component';
 import { useAppSelector } from '../../hooks';
 import ErrorScreen404 from '../error-screen-404/error-screen-404';
 
@@ -32,17 +33,7 @@ function AddReview(): JSX.Element {
               </li>
             </ul>
           </nav>
-
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-              </div>
-            </li>
-            <li className="user-block__item">
-              <a href="/" className="user-block__link">Sign out</a>
-            </li>
-          </ul>
+          <SignOutComponent/>
         </header>
 
         <div className="film-card__poster film-card__poster--small">
