+++ project/src/pages/main-page/main-page.tsx
@@ -1,19 +1,10 @@
-import { Link } from 'react-router-dom';
-import FilmCard from '../../components/film-card/filmCard';
-import LogoFooter from '../../components/logo-footer/logo-footer';
-import Logo from '../../components/logo/logo';
-import GenresList from '../../components/genres-list/genres-list';
+import HeadGuest from '../../components/head-guest/head-guest';
+import MainPageComponent from '../../components/main-page-component/main-page-component';
+import { AuthorizationStatus } from '../../const';
 import { useAppSelector } from '../../hooks';
-import ShowMoreButton from '../../components/show-more-button/show-more-button';
-import { AppRoute, GENRES } from '../../const';
 
 function MainPage(): JSX.Element {
-  const maxRenderedFilmsQuantity = useAppSelector((state) => state.maxRenderedFilmsQuantity);
-  const loadedFilmsState = useAppSelector((state) => state.loadedFilms);
-  const genreState = useAppSelector((state) => state.genre);
-  const filteredFilms = loadedFilmsState.filter((elem) => elem.genre === genreState || genreState === GENRES.ALLGENRES).slice(0, maxRenderedFilmsQuantity);
-  const promoFilm = useAppSelector((state) => state.promoFilm);
-  const { name, released, genre, posterImage, backgroundImage } = promoFilm;
+  const {authorizationStatus} = useAppSelector((state) => state);
   return (
     <div>
       <div className="visually-hidden">
@@ -47,78 +38,8 @@ function MainPage(): JSX.Element {
           </symbol>
         </svg>
       </div>
-
-      <section className="film-card">
-        <div className="film-card__bg">
-          <img src={backgroundImage} alt={name} />
-        </div>
-
-        <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header film-card__head">
-          <Logo />
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <Link to="/mylist">
-                  <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-                </Link>
-              </div>
-            </li>
-            <li className="user-block__item">
-              <a href="/" className="user-block__link">Sign out</a>
-            </li>
-          </ul>
-        </header>
-
-        <div className="film-card__wrap">
-          <div className="film-card__info">
-            <div className="film-card__poster">
-              <img src={posterImage} alt={name} width="218" height="327" />
-            </div>
-
-            <div className="film-card__desc">
-              <h2 className="film-card__title">{name}</h2>
-              <p className="film-card__meta">
-                <span className="film-card__genre">{genre}</span>
-                <span className="film-card__year">{released}</span>
-              </p>
-
-              <div className="film-card__buttons">
-                <button className="btn btn--play film-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list film-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width="19" height="20">
-                    <use xlinkHref="#add"></use>
-                  </svg>
-                  <Link to={AppRoute.MyList}>
-                    <span>My list</span>
-                  </Link>
-                  <span className="film-card__count">9</span>
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section >
-
-      <div className="page-content">
-        <section className="catalog">
-          <h2 className="catalog__title visually-hidden">Catalog</h2>
-          <GenresList />
-          <div className="catalog__films-list">
-            {filteredFilms.map((film, index) => (<FilmCard film={film} key={film.id} />)
-            )}
-          </div>
-          <ShowMoreButton />
-        </section>
-        <LogoFooter />
-      </div>
-    </div >
+      {authorizationStatus === AuthorizationStatus.Auth ? <MainPageComponent/> : <HeadGuest/>}
+    </div>
   );
 }
