+++ project/src/pages/my-list/my-list.tsx
@@ -1,26 +1,20 @@
 import FilmCard from '../../components/film-card/filmCard';
 import LogoFooter from '../../components/logo-footer/logo-footer';
 import Logo from '../../components/logo/logo';
+import SignOutComponent from '../../components/sign-out-component/sign-out-component';
 import { useAppSelector } from '../../hooks';
 
+
 function MyList(): JSX.Element {
-  const loadedFilms = useAppSelector((state) => state.loadedFilms);
+  const { loadedFilms } = useAppSelector((state) => state);
+
   return (
 
     <div className="user-page">
       <header className="page-header user-page__head">
         <Logo />
         <h1 className="page-title user-page__title">My list <span className="user-page__film-count">9</span></h1>
-        <ul className="user-block">
-          <li className="user-block__item">
-            <div className="user-block__avatar">
-              <img src="/img/avatar.jpg" alt="User avatar" width="63" height="63" />
-            </div>
-          </li>
-          <li className="user-block__item">
-            <a href="/" className="user-block__link">Sign out</a>
-          </li>
-        </ul>
+        <SignOutComponent />
       </header>
 
       <section className="catalog">
