+++ project/src/pages/sign-in/sign-in.tsx
@@ -15,6 +15,7 @@ function SignIn(): JSX.Element {
 
   const onSubmit = (authData: AuthData) => {
     dispatch(loginAction(authData));
+    navigate(AppRoute.Root);
   };
 
   const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
@@ -27,6 +28,7 @@ function SignIn(): JSX.Element {
       });
     }
   };
+
   return (
     <div className="user-page">
       <header className="page-header user-page__head">
@@ -47,7 +49,7 @@ function SignIn(): JSX.Element {
             </div>
           </div>
           <div className="sign-in__submit">
-            <button className="sign-in__btn" type="submit" onClick={() => navigate(AppRoute.Root)}>Sign in</button>
+            <button className="sign-in__btn" type="submit">Sign in</button>
           </div>
         </form>
       </div>
